fix: disable transitions in ie11
